Support the environment which does not support OffscreenCanvas

+++ src/renderer/CanvasRenderer/index.ts
@@ -1,5 +1,6 @@
 import EventEmitter from 'eventemitter3'
-import {MalVal, symbolFor as S} from '@/mal/types'
+import CanvasRenderer from './CanvasRenderer'
+import {MalVal} from '@/mal/types'
 import {mat3} from 'gl-matrix'
 
 export interface ViewerSettings {
@@ -7,63 +8,60 @@ export interface ViewerSettings {
 	guideColor?: string
 }
 
-export default class CanvasRender extends EventEmitter {
-	private canvas: HTMLCanvasElement
-	private offscreenCanvas: OffscreenCanvas
-	private worker: Worker
+export default class CanvasRenderDelegate extends EventEmitter {
+	private canvas!: HTMLCanvasElement
+	private worker!: Worker
+	private renderer!: CanvasRenderer
 	private rendering = false
 
-	constructor(canvas: HTMLCanvasElement) {
+	constructor() {
 		super()
-
-		this.canvas = canvas
-		this.offscreenCanvas = this.canvas.transferControlToOffscreen()
-
-		this.worker = new Worker('./worker.ts', {type: 'module'})
-		this.worker.onmessage = e => this.onMessage(e)
-		this.worker.postMessage(
-			{
-				type: 'init',
-				params: {canvas: this.offscreenCanvas}
-			},
-			[this.offscreenCanvas]
-		)
 	}
 
 	public get isRendering() {
 		return this.rendering
 	}
 
-	public resize(...args: number[]) {
-		let width, height, dpi
+	public async init(canvas: HTMLCanvasElement) {
+		this.canvas = canvas
 
-		if (args.length === 0) {
+		if (typeof OffscreenCanvas !== 'undefined') {
+			this.worker = new Worker('./worker.ts', {type: 'module'})
+			this.worker.onmessage = e => this.onMessage(e)
+			const offscreenCanvas = canvas.transferControlToOffscreen()
+			await this.postMessageAndWait('init', {canvas: offscreenCanvas}, [
+				offscreenCanvas
+			])
+		} else {
+			this.renderer = new CanvasRenderer()
+			await this.postMessageAndWait('init', {canvas})
+		}
+	}
+
+	public async resize(width?: number, height?: number, dpi?: number) {
+		if (dpi === undefined) {
 			width = this.canvas.clientWidth
 			height = this.canvas.clientHeight
 			dpi = window.devicePixelRatio || 1
-		} else {
-			;[width, height, dpi] = args
 		}
 
-		this.worker.postMessage({type: 'resize', params: {width, height, dpi}})
-	}
-
-	public dispose() {
-		this.worker.terminate()
+		return await this.postMessageAndWait('resize', {width, height, dpi})
 	}
 
 	public async render(ast: MalVal, settings: ViewerSettings) {
 		this.rendering = true
-
 		await this.postMessageAndWait('render', {ast, settings})
-
 		this.rendering = false
-
-		return
 	}
 
 	public async getImage() {
 		return await this.postMessageAndWait('get-image')
+	}
+
+	public async dispose() {
+		if (this.worker) {
+			this.worker.terminate()
+		}
 	}
 
 	private postMessageAndWait(
@@ -71,15 +69,26 @@ export default class CanvasRender extends EventEmitter {
 		params: any = null,
 		transfer: Transferable[] = []
 	) {
-		return new Promise(resolve => {
-			this.worker.postMessage({type, params}, transfer)
-			this.once(type, (result: any) => {
-				resolve(result)
-			})
+		return new Promise((resolve, reject) => {
+			if (this.worker) {
+				this.worker.postMessage({type, params}, transfer)
+				this.once(type, (result: any) => {
+					if (result instanceof Error) {
+						reject(result)
+					} else {
+						resolve(result)
+					}
+				})
+			} else {
+				this.renderer
+					.postMeessage(type, params)
+					.then(resolve)
+					.catch(reject)
+			}
 		})
 	}
 
-	private onMessage(e: any) {
+	private onMessage(e: MessageEvent) {
 		const {type, params} = e.data
 		this.emit(type, params)
 	}
